web/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
supabase==2.0.3
python-dotenv==1.0.0
pytest==7.4.0
requests==2.31.0